+++ CHANGELOG.md
@@ -21,6 +21,10 @@
         - char* str_trim(const char* text);
         - char* str_to_upper(const char* text);
         - char* str_to_lower(const char* text);
+    - Improved math library
+        - uint64_t abs(const int64_t n);
+        - int64_t clamp(const int64_t n, const int64_t min, const int64_t max);
+        - bool is_power_of_two(const int64_t n);
     - Sorted files in Types folder in subfolders
     - Fixed compiler warninigs